rename event comp for home
