+++ src/lib/components/HomeEvent.svelte
